feat: add upload-routes.js
