+++ api/app.js
@@ -1,7 +1,6 @@
 import cors from "cors"
 import express from "express"
 import helmet from "helmet"
-import fs from "fs"
 import logger from "./utils/logger.js"
 import errorHandler from "./middlewares/error-handler.js"
 import { fileProcessorQueue } from "./queues/file-processor-queue.js"
@@ -9,8 +8,9 @@ import { FlowProducer } from "bullmq"
 import { getAllFilesInDir } from "./utils/file-utils.js"
 import { ExpressAdapter } from "@bull-board/express"
 import { createBullBoard } from "@bull-board/api"
+import uploadRoutes from "./routes/upload-routes.js"
 import { BullMQAdapter } from "@bull-board/api/bullMQAdapter.js"
-import { DEFAULT_JOB_REMOVE_CONFIG } from "./config/config.js"
+import { DEFAULT_JOB_REMOVE_CONFIG, UPLOAD_DIR } from "./config/config.js"
 import { createZipFolderQueue } from "./queues/create-zip-folder-queue.js"
 
 const PORT = process.env.PORT || 8000
@@ -34,213 +34,13 @@ app.use(express.json())
 // Initialize Middleware
 app.use(cors({ origin: "*" }))
 app.use(helmet())
+
+// Initialize Routes
 app.use("/admin/queues", serverAdapter.getRouter())
-
-const dest = "uploads/"
-
-// Helper function for logging
-const logRequestDetails = (req) => {
-  logger.info(`Received ${req.method} request on ${req.url}`)
-  logger.info(`Headers: ${JSON.stringify(req.headers)}`)
-}
-
-app.get("/", (req, res) => {
-  logRequestDetails(req)
-  return res.send("Hello World!")
-})
-
-app.post("/", function (req, res) {
-  logRequestDetails(req)
-  return res.send("Got a POST request")
-})
-
-app.get("/upload/status", (req, res, next) => {
-  logRequestDetails(req)
-
-  const uniqueFileId = String(req.headers["x-file-name"])
-  const uniqueJobId = String(req.headers["x-job-id"])
-  const uploadId = uniqueJobId + "_" + uniqueFileId
-  const fileSize = parseInt(String(req.headers["file-size"]), 10)
-  const jobDir = dest + `${uniqueJobId}/`
-
-  if (!fs.existsSync(dest)) {
-    logger.warn(`Destination folder ${dest} does not exist. Creating...`)
-    fs.mkdirSync(dest)
-  }
-
-  if (!fileSize) {
-    const error = new Error("No file-size header found")
-    error.status = 400
-    return next(error)
-  }
-
-  if (!uniqueFileId) {
-    const error = new Error("No x-file-name header found")
-    error.status = 400
-    return next(error)
-  }
-
-  if (!uniqueJobId) {
-    const error = new Error("No x-job-id header found")
-    error.status = 400
-    return next(error)
-  }
-
-  logger.info(
-    `Checking upload status for file: ${uniqueFileId}, expected fileSize: ${fileSize}`,
-  )
-
-  if (uniqueFileId) {
-    try {
-      const stats = fs.statSync(jobDir + uniqueFileId)
-      if (stats.isFile()) {
-        if (fileSize === stats.size) {
-          logger.info(`File ${uniqueFileId} is already fully uploaded.`)
-          return res.send({
-            status: "ALREADY_UPLOADED_FILE",
-            uploaded: stats.size,
-          })
-        }
-        if (!uploads[uploadId]) uploads[uploadId] = {}
-        uploads[uploadId]["bytesReceived"] = stats.size
-        logger.info(`File ${uniqueFileId} has ${stats.size} bytes uploaded.`)
-        return res.send({ uploaded: stats.size })
-      }
-    } catch (err) {
-      const upload = uploads[uploadId]
-      if (upload) {
-        logger.info(
-          `Resuming file upload for ${uniqueFileId}, uploaded: ${upload.bytesReceived}`,
-        )
-        return res.send({
-          uploaded: upload.bytesReceived,
-          status: "RESUMED_FILE",
-        })
-      } else {
-        logger.info(`New file upload initiated for ${uniqueFileId}`)
-        return res.send({ uploaded: 0, status: "NEW_FILE" })
-      }
-    }
-  }
-})
-
-let uploads = {}
-
-app.post("/upload/files", (req, res, next) => {
-  logRequestDetails(req)
-
-  const uniqueFileId = String(req.headers["x-file-name"])
-  const uniqueJobId = String(req.headers["x-job-id"])
-  const uploadId = uniqueJobId + "_" + uniqueFileId
-  const match = req.headers["content-range"].match(/(\d+)-(\d+)\/(\d+)/)
-  const start = parseInt(match[1])
-  const fileSize = parseInt(String(req.headers["file-size"]), 10)
-
-  const jobDir = dest + `${uniqueJobId}`
-
-  logger.info(
-    `Received file upload request for ${uploadId}, job id is ${uniqueJobId}`,
-  )
-
-  if (!fs.existsSync(jobDir)) {
-    logger.warn(`Destination folder ${dest} does not exist. Creating...`)
-    fs.mkdirSync(jobDir, { recursive: true })
-  }
-
-  if (uploads[uploadId] && fileSize === uploads[uploadId].bytesReceived) {
-    logger.info(
-      `File ${uniqueFileId} already uploaded completely with the job ${uniqueJobId}`,
-    )
-    return res.status(200).send("File already uploaded")
-  }
-
-  if (!uniqueFileId) {
-    const error = new Error(`No x-file-name header found`)
-    error.status = 400
-    return next(error)
-  }
-
-  if (!uniqueJobId) {
-    const error = new Error("No x-job-id header found")
-    error.status = 400
-    return next(error)
-  }
-
-  if (!uploads[uploadId]) uploads[uploadId] = {}
-  const upload = uploads[uploadId]
-
-  let fileStream
-
-  if (!start) {
-    upload.bytesReceived = 0
-    logger.info(`Starting new file upload for ${uploadId}`)
-    fileStream = fs.createWriteStream(`./${jobDir}/${uniqueFileId}`, {
-      flags: "w",
-    })
-  } else {
-    if (upload.bytesReceived != start) {
-      logger.error(
-        `Wrong start byte for ${uniqueFileId}. Expected: ${upload.bytesReceived}, received: ${start}`,
-      )
-      res.writeHead(400, "Wrong start byte")
-      res.end(upload.bytesReceived)
-      return
-    }
-    logger.info(`Resuming file upload for ${uploadId} from byte ${start}`)
-    fileStream = fs.createWriteStream(`./${jobDir}/${uniqueFileId}`, {
-      flags: "a",
-    })
-  }
-
-  req.on("data", function (data) {
-    upload.bytesReceived += data.length
-  })
-
-  req.pipe(fileStream)
-
-  // when the request is finished, and all its data is written
-  fileStream.on("close", function () {
-    logger.info(`File upload complete for ${uploadId}`)
-    return res.status(201).send({ status: "UPLOAD_COMPLETE" })
-  })
-
-  // in case of I/O error - finish the request
-  fileStream.on("error", function (_err) {
-    const error = new Error(
-      `File upload error for ${uploadId}: ${_err.message}`,
-    )
-    error.status = 500
-    return next(error)
-  })
-})
-
-app.post("/upload/complete", (req, res, next) => {
-  logRequestDetails(req)
-
-  const uniqueFileId = String(req.headers["x-file-name"])
-  const uniqueJobId = String(req.headers["x-job-id"])
-
-  if (!uniqueFileId) {
-    const error = new Error(`No x-file-name header found`)
-    error.status = 400
-    return next(error)
-  }
-
-  if (!uniqueJobId) {
-    const error = new Error("No x-job-id header found")
-    error.status = 400
-    return next(error)
-  }
-
-  const uploadId = uniqueJobId + "_" + uniqueFileId
-  delete uploads[uploadId]
-
-  logger.info(`File upload process completed for ${uploadId}`)
-  return res.status(201).send({ status: "SUCCESSFULLY_UPLOADED" })
-})
+app.use("/", uploadRoutes)
 
 app.post("/convert/files", async (req, res, next) => {
-  const uniqueJobId = String(req.headers["x-job-id"])
+  const uniqueJobId = req.headers["x-job-id"]
 
   if (!uniqueJobId) {
     const error = new Error("No x-job-id header found")
@@ -248,7 +48,7 @@ app.post("/convert/files", async (req, res, next) => {
     return next(error)
   }
 
-  const filesDir = dest + `${uniqueJobId}`
+  const filesDir = UPLOAD_DIR + `${uniqueJobId}`
 
   await getAllFilesInDir(filesDir)
     .then(async (files) => {
